requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.40
psycopg[binary]==3.2.6
pydantic[email]==2.11.3
pydantic-settings==2.8.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
httpx==0.28.1
python-multipart==0.0.20

